backend/requirements.txt
# Milestone-4 Dependencies
# Install with: pip install -r backend/requirements.txt

# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
mysql-connector-python==8.2.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Existing dependencies
requests==2.31.0
